requirements.txt
numpy
imutils
sklearn
tensorflow
keras
pillow
opencv-python
